+++ internal/api/docs.go
@@ -1,8 +1,13 @@
 package api
 
 import (
+	"bytes"
+	"text/template"
+
 	"github.com/benthosdev/benthos/v4/internal/docs"
 	httpdocs "github.com/benthosdev/benthos/v4/internal/http/docs"
+
+	_ "embed"
 )
 
 // Spec returns a field spec for the API configuration fields.
@@ -19,13 +24,52 @@ func Spec() docs.FieldSpecs {
 		docs.FieldString("cert_file", "An optional certificate file for enabling TLS.").Advanced().HasDefault(""),
 		docs.FieldString("key_file", "An optional key file for enabling TLS.").Advanced().HasDefault(""),
 		httpdocs.ServerCORSFieldSpec(),
-		docs.FieldObject("basic_auth", "Allows you to specify basic authentication.").WithChildren(
-			docs.FieldBool("enabled", "Whether to use basic authentication in requests.").HasDefault(false),
-			docs.FieldString("username", "Username required to authenticate.").HasDefault(""),
-			docs.FieldString("password_hash", "Hashed password required to authenticate. (base64 encoded)").HasDefault(""),
-			docs.FieldString("realm", "realm of the protection space.").HasDefault("restricted"),
-			docs.FieldString("algorithm", "Encryption algorithm used to generate password_hash.", "md5", "sha256", "bcrypt", "scrypt").HasDefault("sha256"),
-			docs.FieldString("salt", "Salt for scrypt algorithm. (base64 encoded)").HasDefault(""),
-		).Advanced(),
+		httpdocs.BasicAuthFieldSpec(),
 	}
+}
+
+//go:embed docs.md
+var httpDocs string
+
+type templateContext struct {
+	Fields         []docs.FieldSpecCtx
+	CommonConfig   string
+	AdvancedConfig string
+}
+
+// DocsMarkdown returns a markdown document for the http documentation.
+func DocsMarkdown() ([]byte, error) {
+	httpDocsTemplate := docs.FieldsTemplate(false) + httpDocs
+
+	var buf bytes.Buffer
+	err := template.Must(template.New("http").Parse(httpDocsTemplate)).Execute(&buf, templateContext{
+		Fields: docs.FieldObject("", "").WithChildren(Spec()...).FlattenChildrenForDocs(),
+		CommonConfig: `
+http:
+  address: 0.0.0.0:4195
+  enabled: true
+  root_path: /benthos
+  debug_endpoints: false
+`,
+		AdvancedConfig: `
+http:
+  address: 0.0.0.0:4195
+  enabled: true
+  root_path: /benthos
+  debug_endpoints: false
+  cert_file: ""
+  key_file: ""
+  cors:
+    enabled: false
+    allowed_origins: []
+  basic_auth:
+    enabled: false
+    username: ""
+    password_hash: ""
+    algorithm: "sha256"
+    salt: ""
+`,
+	})
+
+	return buf.Bytes(), err
 }