+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 
 ## Unreleased
 
+### Added
+
+- The service-wide `http` config now supports basic authentication.
+
 ## 4.3.0 - 2022-06-23
 
 ### Added